requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==1.1.2
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
bcrypt==3.1.0
billiard
bleach==1.5.0
boto==2.40.0
cffi==1.8.2
channels==1.1.8
channels-api==0.4.0
constantly==15.1.0
coreapi==2.0.0
daphne==1.4.2
dj-database-url==0.4.1
Django==1.10
django-braces==1.8.1
django-cors-middleware==1.3.1
django-debug-toolbar==1.5
django-environ==0.4.0
django-extended-choices==1.0.4
django-extensions==1.7.4
django-extra-fields==0.9
django-filter==0.11.0
django-nose==1.4.4
django-oauth-toolkit==0.11.0
django-rest-swagger==2.0.5
django-storages==1.4.1
djangorestframework==3.4.6
drf-extensions==0.2.7
factory-boy==2.8.1
Faker==0.7.10
funcsigs==1.0.2
future==0.15.0
gunicorn==19.4.5
html5lib==0.9999999
hyperlink==17.3.1
incremental==17.5.0
ipaddress==1.0.18
itypes==1.1.0
jsonfield==1.0.3
kombu
libmagic==1.0
mock==2.0.0
newrelic==2.68.0.50
nose==1.3.7
oauthlib==1.1.2
openapi-codec==1.0.0
pbr==1.10.0
Pillow==4.0
pinax-eventlog==1.1.2
psycopg2==2.7
pycparser==2.14
Pygments==2.0.2
python-dateutil==2.6.0
python-magic==0.4.10
pytz==2016.4
redgreenunittest==0.1.1
requests==2.9.1
simplejson==3.8.2
six==1.10.0
sqlparse==0.2.1
tblib==1.3.0
Twisted==17.9.0
txaio==2.8.2
Unidecode==0.4.19
uritemplate==3.0.0
whitenoise==3.3.1
zope.interface==4.4.3
django-background-tasks==1.1.13
imutils==0.4.6
numpy
keras
tensorflow
pygraphviz
celery                  # celery and redis are for queuing tasks
redis
